修改一些Bug.增加比例布局。

+++ DuiLib/Core/UIRender.cpp
@@ -1980,7 +1980,7 @@ SIZE CRenderEngine::GetTextSize( HDC hDC, CPaintManagerUI* pManager , LPCTSTR ps
 	if( pstrText == NULL || pManager == NULL ) return size;
 	::SetBkMode(hDC, TRANSPARENT);
 	HFONT hOldFont = (HFONT)::SelectObject(hDC, pManager->GetFont(iFont));
-	GetTextExtentPoint32(hDC, pstrText, _tcslen(pstrText) , &size);
+	GetTextExtentPoint32(hDC, pstrText, (int)_tcslen(pstrText) , &size);
 	::SelectObject(hDC, hOldFont);
 	return size;
 }
